# apps/docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UltraContext",
    "description": "Manage conversation contexts with automatic versioning. Every mutation creates a new branch, preserving full history. All API-key-authenticated requests are rate-limited to protect against abuse — a 429 carries a `RateLimitError` body with a `retry_after_sec` field. Every response carries an `X-Request-Id` header for correlation with server logs.",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://api.ultracontext.ai"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/contexts": {
      "post": {
        "operationId": "createContext",
        "summary": "Create context",
        "description": "Create a new context. Optionally copy from an existing context using the `from` parameter, and specify a `version` to fork from a specific version.",
        "x-mint": {
          "content": "<Tip>Use `from` with `version`, `at`, or `before` to fork from a specific point in history. Learn more in the [Fork & Clone Contexts](/guides/fork-clone-contexts) guide.</Tip>"
        },
        "x-codeSamples": [
          {
            "lang": "typescript",
            "label": "TypeScript",
            "source": "const ctx = await uc.create()\n\n// With metadata\nconst ctx = await uc.create({ metadata: { name: 'my-session' } })\n\n// Copy from existing (latest version)\nconst fork = await uc.create({ from: 'ctx_abc123' })\n\n// Fork from specific version\nconst v1Fork = await uc.create({ from: 'ctx_abc123', version: 1 })\n\n// Fork at specific message index\nconst partial = await uc.create({ from: 'ctx_abc123', at: 5 })\n\n// Fork from point-in-time\nconst snapshot = await uc.create({ from: 'ctx_abc123', before: '2024-01-15T10:30:00Z' })"
          },
          {
            "lang": "python",
            "label": "Python",
            "source": "ctx = uc.create()\n\n# With metadata\nctx = uc.create(metadata={'name': 'my-session'})\n\n# Copy from existing (latest version)\nfork = uc.create(from_='ctx_abc123')\n\n# Fork from specific version\nv1_fork = uc.create(from_='ctx_abc123', version=1)\n\n# Fork at specific message index\npartial = uc.create(from_='ctx_abc123', at=5)\n\n# Fork from point-in-time\nsnapshot = uc.create(from_='ctx_abc123', before='2024-01-15T10:30:00Z')"
          },
          {
            "lang": "bash",
            "label": "cURL",
            "source": "curl -X POST https://api.ultracontext.ai/contexts \\\n  -H \"Authorization: Bearer $API_KEY\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"metadata\": {\"name\": \"my-session\"}}'\n\n# Fork from specific version\ncurl -X POST https://api.ultracontext.ai/contexts \\\n  -H \"Authorization: Bearer $API_KEY\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"from\": \"ctx_abc123\", \"version\": 1}'\n\n# Fork from point-in-time\ncurl -X POST https://api.ultracontext.ai/contexts \\\n  -H \"Authorization: Bearer $API_KEY\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"from\": \"ctx_abc123\", \"before\": \"2024-01-15T10:30:00Z\"}'"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContextInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Context created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Context"
                }
              }
            }
          },
          "400": {
            "description": "Invalid index (when using `at` with `from`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Version not found (when using `version` with `from`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listContexts",
        "summary": "List contexts",
        "description": "List all root contexts for your project. Sorted by `created_at` descending (newest first).",
        "x-codeSamples": [
          {
            "lang": "typescript",
            "label": "TypeScript",
            "source": "// List all contexts\nconst { data } = await uc.get()\n\n// With limit\nconst { data } = await uc.get({ limit: 10 })\n\n// Filter by source\nconst { data } = await uc.get({ source: 'claude' })\n\n// Filter by project\nconst { data } = await uc.get({ project_path: '/Users/alice/Code/myapp' })\n\n// Combine filters\nconst { data } = await uc.get({ source: 'codex', host: 'Mac-mini.local', after: '2026-03-01T00:00:00Z' })"
          },
          {
            "lang": "python",
            "label": "Python",
            "source": "# List all contexts\nresult = uc.get()\n\n# With limit\nresult = uc.get(limit=10)\n\n# Filter by source\nresult = uc.get(source='claude')\n\n# Filter by project\nresult = uc.get(project_path='/Users/alice/Code/myapp')\n\n# Combine filters\nresult = uc.get(source='codex', host='Mac-mini.local', after='2026-03-01T00:00:00Z')"
          },
          {
            "lang": "bash",
            "label": "cURL",
            "source": "curl https://api.ultracontext.ai/contexts \\\n  -H \"Authorization: Bearer $API_KEY\"\n\n# With limit\ncurl \"https://api.ultracontext.ai/contexts?limit=10\" \\\n  -H \"Authorization: Bearer $API_KEY\"\n\n# Filter by source + project\ncurl \"https://api.ultracontext.ai/contexts?source=claude&project_path=/Users/alice/Code/myapp\" \\\n  -H \"Authorization: Bearer $API_KEY\"\n\n# Time range\ncurl \"https://api.ultracontext.ai/contexts?after=2026-03-01T00:00:00Z&before=2026-03-08T00:00:00Z\" \\\n  -H \"Authorization: Bearer $API_KEY\""
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of contexts to return",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by agent source (e.g. `claude`, `codex`, `openclaw`)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter by user identifier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host",
            "in": "query",
            "description": "Filter by machine hostname",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project_path",
            "in": "query",
            "description": "Filter by project directory path (e.g. `/Users/alice/Code/myapp`)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "description": "Filter by session identifier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Only contexts created after this ISO 8601 timestamp",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Only contexts created before this ISO 8601 timestamp",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of contexts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Context"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid limit (must be a positive integer)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/contexts/{id}/branches/{name}": {
      "delete": {
        "operationId": "deleteBranch",
        "summary": "Delete a branch",
        "description": "Remove a branch name. This deletes the POINTER only — version data is never touched, so the version it pointed at stays readable by its id. Unpinning a name can never destroy history.",
        "x-codeSamples": [
          {
            "lang": "typescript",
            "label": "TypeScript",
            "source": "const { deleted, name } = await uc.deleteBranch('ctx_abc123', 'release-1.2')"
          },
          {
            "lang": "python",
            "label": "Python",
            "source": "result = uc.delete_branch('ctx_abc123', 'release-1.2')"
          },
          {
            "lang": "bash",
            "label": "cURL",
            "source": "curl -X DELETE https://api.ultracontext.ai/contexts/ctx_abc123/branches/release-1.2 \\\n  -H \"Authorization: Bearer $API_KEY\""
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Context ID (ctx_...)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Branch name (same rules as when creating it)",
            "schema": {
              "type": "string",
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Branch name removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "example": true
                    },
                    "name": {
                      "type": "string",
                      "example": "release-1.2"
                    }
                  },
                  "required": [
                    "deleted",
                    "name"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid branch name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Context not found, or no branch with that name on this context",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from your UltraContext dashboard. The /v1/keys routes additionally require the admin key."
      }
    },
    "schemas": {
      "Context": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier (ctx_...)",
            "example": "ctx_a1b2c3d4e5f6"
          },
          "metadata": {
            "type": "object",
            "description": "User-defined metadata",
            "additionalProperties": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp"
          }
        }
      },
...
      "CreateContextInput": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "description": "Source context ID to copy from (optional)"
          },
          "version": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "description": "When forking with `from`, copy from this version — either its immutable id (`ctx_…`, preferred) or its positional index (0-indexed, deprecated alias). Negative indexes count back from the head, so -1 is the latest. An unknown id is a 404; a fractional number is a 400."
          },
          "at": {
            "type": "integer",
            "description": "When forking with `from`, copy only messages 0 through this index"
          },
          "before": {
            "type": "string",
            "format": "date-time",
            "description": "When forking with `from`, copy point-in-time state before this ISO timestamp"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "User-defined metadata for this context"
          }
        },
        "description": "Create a new context with optional fork params and metadata"
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "enum": [
              "not_found",
              "invalid_input",
              "conflict",
              "internal"
            ],
            "description": "Machine-readable error code. `conflict` is retryable — the response also carries a `Retry-After` header; wait that many seconds and retry the request verbatim."
          }
        }
      }
    }
  }
}